requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==4.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==21.1
django-phonenumber-field==6.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
phonenumbers==8.12.47
Pillow==9.1.0
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
wincertstore==0.2
